[dagster-airlift] fix tests
