+++ examples/experimental/dagster-airlift/conftest.py
@@ -1,0 +1,1 @@
+pytest_plugins = ["dagster_airlift.shared_fixtures"]
