+++ examples/experimental/dagster-airlift/examples/peering-with-dbt/peering_with_dbt_tests/conftest.py
@@ -1,7 +1,6 @@
 import os
-import signal
 import subprocess
-from typing import Any, Generator
+from typing import Generator
 
 import pytest
 from dagster._core.test_utils import environ
@@ -19,21 +18,3 @@ def setup_fixture() -> Generator[None, None, None]:
     ):
         yield
     subprocess.run(["make", "wipe"], check=True)
-
-
-@pytest.fixture(name="airflow_instance")
-def airflow_instance_fixture(setup: None) -> Generator[Any, None, None]:
-    process = subprocess.Popen(
-        ["airflow", "standalone"],
-        env=os.environ,  # since we have some temp vars in the env right now
-        shell=False,
-        stdout=subprocess.PIPE,
-        preexec_fn=os.setsid,  # noqa  # fuck it we ball
-    )
-    assert process.stdout is not None
-    for line in process.stdout:
-        if "Airflow is ready" in line.decode():
-            break
-    yield process
-    # Kill process group, since process.kill and process.terminate do not work.
-    os.killpg(process.pid, signal.SIGKILL)
